HREPTWO-22 Experimental implementation of dynamic gui loading version 1
In the upper right corner of the gui there is a new link 'Workflow'
Clicking on this link results in a pop-up dialog

The content (implementation: java code + markup html) of this dialog is dependendent on the selected node in the tree.

Walkthrough:
1>create a node (anywhere in the tree)
2>add a property 'renderer' with value 'org.hippocms.repository.frontend.workflow.foo.FooDialog'
3>select the node
4>click on the workflow link
=>Dialog 'FooDialog' pops up

5>change the value to  'org.hippocms.repository.frontend.workflow.bar.BarDialog'
6>select the node
7>click on the workflow link
=>Dialog 'BarDialog' pops up

Offcourse FooDialog and BarDialog are sample implementations and part of the frondend module,

However this also works for external modules, just drop a jar containing your implementation
(org.example.MyVeryImportantWorkflowDialog) into the webapp classpath and set the renderer property
accordingly.



  


svn path=/hippo-repository-2/trunk/; revision=7379