HREPTWO-194: automatically detect changes in the configuration/initialize tree
When properties, used by repository initialization, are added to the configuration/initialize
tree these initialization steps are carried out at runtime.  So it is possible to configure
the repository while it is running!

Note that nodes that have been created as a result of earlier initialization will not be
deleted from the repository.  It is only possible to add content.


svn path=/hippo-ecm/trunk/; revision=9071