HREPTWO-250: use multiple editor tabs.
The node model is no longer implicitly updated by Plugin, as this disturbs the simple "forward to children" philosophy.
There is still a hack in this solution, as the different tabs must have different plugin-ids.  However, the plugin factory uses the id to look up the definition so that has to be the same 
for each editor perspective.

Some parameters have been introduced for the perspectives, making the title configurable.

For actual testing, prototype support has been added to the template.  At least all the mandatory fields should be defined, to obtain valid nodes.


svn path=/hippo-ecm/trunk/; revision=9860