HREPTWO-337: support same-name siblings.
When the template node (of type hippo:template) has the "hippo:multiple" property set to true, it is now possible to dynamically add and 
remove nodes.  The models that describe the siblings are produced by the MultiProvider class and shown by the MultiTemplate, using the 
MultiView.  Models that correspond to fields, independent of how often they (can) appear, are produced by the FieldProvider.  The 
FieldModel class contains the node that a template applies to and a descriptor that identifies the field.


svn path=/hippo-ecm/trunk/; revision=9955