HREPTWO-250: introduce IPluginModel
This model unifies the Plugin constructor, model and the message.  It extends the
IModel and is e.g. implemented by JcrNodeModel.  This model extends the Wicket
IModel interface with an extra method, getMapRepresentation.

Since JcrNodeModel now has a constructor that takes an IPluginModel as the
(only) argument, it is straightforward to create and interpret messages.

The interface is experimental. 

svn path=/hippo-ecm/trunk/; revision=10051