HREPTWO-444, HREPTWO-337: restructure templates and export to .cnd
The Cnd class prints the templates that are defined for an application.
The arguments are <application> (e.g. hippo:console) and, optionally, a path to the repository 
(e.g. rmi://localhost:1099/hipporepository).
The JcrSessionModel has been liberated from the UserSession, to reduce the usage of 
Session.get().  (one should not have to invent a Request object to access the repository)

Templates have been restructured with a stricter separation between fields and templates, 
trivializing the node type extraction.


svn path=/hippo-ecm/trunk/; revision=10274