HREPTWO-250: experiment with new architecture
Initial version of proposed architecture.  It still uses the existing
architecture for support (jcr session etc), but instead uses a distributed
services model similar to OSGi.  A minimal application consisting of a
browser and an editor is implemented.  These two do not yet communicate, 
so it is not yet actually possible to browse...

Usage: add the Adapter as an old-style plugin to the configuration.


svn path=/hippo-ecm/trunk/; revision=11115