HREPTWO-250: communicating plugins
Plugins can now share models.  The ModelReference uses a TopicService to communicate with peers.
The ModelReference updates the RenderService when a SET_MODEL message is received.

A custom target, the PluginRequestTarget, is used instead of the normal AjaxRequestTarget.
This removes many steps from the process of handling ajax events.  A RenderService is always
added to the target when its model changes.  The target handles the filtering of updates. 

svn path=/hippo-ecm/trunk/; revision=11133