HREPTWO-250: Multiple editors are now supported by the tabs plugin.
The tabs plugin itself no longer has knowledge of editing; it only listens for
services that register under the TabsPlugin.TAB_ID service ID.  Closing tabs
(IDynamicService) is not yet supported.

The MultiEditorPlugin listens on the model topic for node selection events.
When a node is selected, it instantiates a new editor when there is no editor
yet for it.

svn path=/hippo-ecm/trunk/; revision=11145