HREPTWO-250: decorate editors
Introduced decorators.  These live under the decorator Id that is provided by an IRenderService.
Containers can use these decorators to display additional information about a renderer.
The TabsPlugin uses the ITitleDecorator that is provided by the EditorPlugin to customize the
title of the tab.

Render services now obtain their wicketId from their parent in a separate bind() call.  This should
make it possible for the parent to wrap a child renderer, e.g. to decorate it with a title.

The PluginRequestTarget has been simplified.  It overrides the addComponent method, so that any
updates by child components are properly intercepted.

Also, a start has been made to provide dialog support.  This is untested at the moment.  The idea
is that there will be one IDialogService on a page, that can then be used by any renderer to display
a popup.  This should even make it possible for hidden plugins to pop up alerts.

svn path=/hippo-ecm/trunk/; revision=11149