HREPTWO-767 HREPTWO-768: atomize remodel workflow
The remodel workflow operation now is an "all or nothing" operation.  Mandatory fields are copied from
the prototype, if one is available.  Child nodes are now copied as well as properties.

To be able to rollback changes to the namespace registry, a subclass of NamespaceRegistryImpl is
used.  While it is "open", namespace registration is temporary until "committed".
When the remodel operation fails, the namespace mapping can thus be reverted.


svn path=/hippo-ecm/trunk/; revision=11234