HREPTWO-250: simplify services
Each service now has a optional, unique service ID.  Perspectives implement the IViewService.
These can be used for the editor and browser perspectives.

A basic idea starts to form to move "application flow" into the frontend workflow plugins.  Whether
these use any repository workflow is not relevant.  Workflows communicate with services that are
exposed by the rest of the application.  Those services should be as dumb as possible.

Other (render) services communicate in a peer-to-peer fashion.  For this communication, topics
are used.  The primary use at the moment is communicating model updates.

svn path=/hippo-ecm/trunk/; revision=11349