HREPTWO-855: revert TabbedPanel to wrap perspectives in <div>
Using a bare <table>, expecting all child plugins to wrap themselves in <div>, does not work with general plugins.

It might be possible to implement a plugin hierarchy by using <table> tags instead of <div>, but then all plugins
should be converted.  We'll end up with lots of nested tables.


svn path=/hippo-ecm/trunk/; revision=11362