HREPTWO-273: support variations
The "variation" parameter describes what variation of the markup should be used.  For a plugin of class MyFancyPlugin with variation "slick", the
markup MyFancyPlugin_slick.html will be used.

(the dutch "vet" could of course be used as well, but I'm not sure if english readers would associate veterinarians with fancy interfaces)

svn path=/hippo-ecm/trunk/; revision=11363