HREPTWO-816: support multiple template editor modes
This is a partial fix.  Only one mode is currently recognised, "edit".  The mode is propagated to nested templates and can be set in a
"field" item node in the template.  So it is possible to embed a read-only template for a particular field.

Ideally, it would be possible to define different templates altogether for different modes.  These would show the same node in different ways,
not just edit and view with the same layout.

svn path=/hippo-ecm/trunk/; revision=11371