HREPTWO-250: convert config to Map interface
The IValueMap interface (from Wicket) is reused, extended with the possible to store objects of the extended type in the map itself.

The JcrPluginConfig implementation of the new interface is untested as of yet.  It should provide an alternative to the old PluginDescriptor,
but then it's both read- and writeable.

svn path=/hippo-ecm/trunk/; revision=11381