HREPTWO-250: organise plugins into clusters
Clusters are basically templates.  A group of plugins that is started at the same
time.  They can be configured as a single plugin; the constituent plugins can get
at the shared configuration by using "template:..." values.

The ServiceTracker has been simplified and promoted into the primary method of
listening for service registrations.  It no longer keeps a list of services, as
this has been moved into the IPluginContext.

A separate IPluginControl controls the lifecycle of the plugin.  It only has one
method, stopPlugin().

It may be possible to move some of the cleanup tasks into the framework by
intercepting them in the PluginContext.


svn path=/hippo-ecm/trunk/; revision=11484