HREPTWO-410: add workflow to edit perspective
The reviewed-actions workflow has been added to the edit perspective.  This perspective forms its own cluster, so it can
be dynamically created by the multi-edit plugin.

The cluster context is wrapped by a decorator when the config service is obtained from the config factory.
This is necessary to be able to be able to define unique names for services running in the cluster.  (two instances don't interfere)
It is possible to declare local names by using the "{cluster}" prefix in the value.  Use the "cluster:" prefix in individual plugin configs
in the cluster to refer to cluster-wide properties.

svn path=/hippo-ecm/trunk/; revision=11734