HREPTWO-670: update plugin architecture description
it's a bit late, but hopefully you can still take a look at it, Niels.

The new architecture is described in quite some detail, so I may want to
cut some out of it in favor of a more pragmatic approach.  I've left the old
documentation in place, more as a hint to myself than as anything I want to
leave in there.


svn path=/hippo-ecm/trunk/; revision=11914