HREPTWO-1103: different content model for templatetype
There are some unrelated changes as well, but am pressed for time.

The folder workflow at the moment has knowledge about the model; this should change.

The hipponamespaceregistry has gone; namespaces are directly registered with JR.  Rollback is
out of the window as we want to save every node that is converted.  Users should make backups.

Namespace versioning is carried out in the frontend.  A user should (eventually) be able to
state the exact URI she wants to use.



svn path=/hippo-ecm/trunk/; revision=12470