HREPTWO-1362: support validation and feedback
Render service feedback (info/warn/error/debug) is redirected to a feedback panel.  This is achieved by registering an IFeedbackMessageFilter with the wicket.feedback configuration value.  Feedback panels pick up the messages by using the registered filters.  They should always be added to the request target by the render service.

Validation services should be invoked before executing any workflow steps.  Support for this has been added to the AbstractWorkflowPlugin.  Validation always occurs for WorkflowActions that are added with addWorkflowAction.  The EditorPlugin registers its form as a validation service, so that any IFormValidators or IValidators that are added to (form)components will be evaluated before a workflow step is executed.

Finally, the TemplateEditingWorkflowPlugin provides a validation service that verifies that no two fields have the same path and uses the feedback mechanism to inform the user.

svn path=/hippo-ecm/trunk/; revision=12951