HREPTWO-1104: cannot reproduce issue consistently; althrough initially the memory will run up, it levels off.  Since using RMI there has been no memory leak this issue can be deemed not important enough, if present.
The committed test is not a true unit test: it will not fail when enabled
when there is a memory leak.  It will print out the memory usage, where
a human can interpret a steadily upgoing number as a memory leak.
To run this test; enable one of the ignored test by replacing @Ignore
with @Test.
To test with rmi, then do this, but also add the test to the RemoteTest
in the repository/test package and run it from there.


svn path=/hippo-ecm/trunk/; revision=12983