HREPTWO-1447: create document and folder model services
The more tricky parts of model updates have been moved to the browser perspective.  It is responsible for keeping the models in sync and selecting a folder when it is set as the document model.

A new method has been added to the JcrItemModel; hasAncestor.  It is a convenience method to safely (no NPEs) determine whether one model is the (JCR) ancestor of the other.

svn path=/hippo-ecm/trunk/; revision=13107