HREPTWO-1448: pluggable yui markup
Introducing the "yui:id" attribute for identifying yui elements in markup.  The values for these are used in the config and (as css classes) in the stylesheet.
The client enhances these elements with an element ID of the form parent-markup-id ":" yui-id, so that they can be used by other (yui) code.
These ids are therefore under the control of Wicket, as the parent-markup-id will typically be a wicket markup-id.  When we have stable wicket markup ids, the yui ids will also be stable for (selenium) testing.

Unfortunately, CSS3 selectors of the form div[yui|id=...] are not yet supported by major browsers.  A workaround is put in place by adding the yui:id attributes as a css class.

The actual problem has been fixed by using an empty panel when the number of tabs drops to zero.

svn path=/hippo-ecm/trunk/; revision=13183