HREPTWO-1513: allow author to create documents
The name is checked in the frontend, i.e. OK is disabled until name is non-null.
The type hippostd:directory has been made writable for authors.  The derived data engine
checkPermission has been removed, as it threw an access control exception on a newly
created node; the access manager should handle this when setProperty is called.
The workflow manager didn't catch the AccessControlException when iterating over roles.

svn path=/hippo-ecm/trunk/; revision=13728