HREPTWO-1851, HREPTWO-1852: i18n from repository
Initial checkin for providing translations from the repository.  Components can intercept the
Localizer by implementing the IStringResourceProvider interface.  The AbstractRenderService
does this by default, delegating the retrieval to an IModelProvider<String>.  This provider
uses a set of keys to locate an appropriate model.  The i18n impl JcrSearchingProvider uses
a simple search, filter, sort algorithm to find mathing nodes.

The way titles, labels and other strings are passed has changed from String to IModel.  This is
more in line with the way treats these things internally.  It also seems the natural way to make
it possible to create an interface for the CMS Translator, as the model can have the
intelligence needed to find a suitable storage place when setObject() is called.

A lot of properties files have been created for workflow plugins.  There are likely to be some
missing, but it should be a good start.  These will be supported independent of the final
solution for repository storage of translations.

svn path=/hippo-ecm/trunk/; revision=14379