HREPTWO-499: templates, user management, folder workflow
The type descriptions have been moved into the frontend engine.  Queries
have been renamed back to 'New Document', 'New Folder', to maintain
compatibility with M7.

Templates can be translated, as can the names (captions) for fields.  There
is no UI yet to do this, though.  The template format as edited with the
builder is unchanged.

svn path=/hippo-ecm/trunk/; revision=14585