HREPTWO-1935: use plugins in xinha dialogs
Extracted the folder views from the browser perspective, into the abstract browse view.
The browse service abstracts the "browser" service, making available two models
(folder and document) for other plugins.

The xinha image picker uses this mechanism to start a custom cluster for the dialog.
The BrowserPlugin in this cluster uses the existing folder views to display the gallery.

The IPluginConfigService has acquired an additional method to list the clusters of a
cluster folder.  This takes away the responsibility of creating cluster ids from the
plugins that want to instantiate a cluster.


svn path=/hippo-ecm/trunk/; revision=15233