HREPTWO-1466: plugins in dialogs
The standard dialog service is now rendered as part of the ajax request,
dialogs should use the render() callback to render plugins that they've
embedded.

The ok() and cancel() template methods have been renamed to onOk and
onCancel(), to conform to the standard (wicket) naming convention.  The
automatic handling of exceptions has gone, dialogs should handle these
themselves.  (if you find that you need this auto-handling, create a base class
that subclasses the AbstractDialog and provides another template method)

It is now possible to specify the initial dimensions for a dialog in the
dialog itself.  For this, put integer values for "width" and "height" in
the getProperties ivaluemap.

svn path=/hippo-ecm/trunk/; revision=15635