HREPTWO-1237: use standard dialog service
Removed the xinha specific dialog infrastructure.  The dialogs themselves are
no longer full plugins, but are now AbstractDialog subclasses.  The clusters in
the cms-pickers category now have the well-defined goal of picking documents.

The base model that is now used for the dialogs is a map with some additional
methods of a more persistence API kind.  (we may want to seperate this API from
the (Map) object, maybe in a separate type of Model)

svn path=/hippo-ecm/trunk/; revision=15636