HREPTWO-2052: use plugins in link picker dialog
The same clusters that have been defined for the xinha pickers are reused.  The
LookupDialog has been moved to the frontend-plugins project as it hasn't turned
out to be sufficiently generic to be in the engine.

The plugins in the picker clusters were moved to the addon-standards project.

The HTML cleaner has been moved into a separate service, eliminating
duplication of configuration.  This makes the Xinha configuration a lot
smaller, so it's now reasonable to customize the Xinha plugin by copying the
basic configuration as used in the system and hippostd templates.

Variable expansion (e.g. use of ${cluster.id} in templates) has been
implemented for nested configs.  (necessary to be able to give the correct
model to the loaded cluster)

svn path=/hippo-ecm/trunk/; revision=15716