HREPTWO-2145: add indicator and suppress errors
The indicator hopefully makes sure that the cms is actually doing something,
while the error suppression makes sure that there won't be an internal error
page.

The actual problem is, of course, that the client can send requests to
components that have been removed in other requests.  But in wicket-ajax?


svn path=/hippo-ecm/trunk/; revision=15897