HREPTWO-2205: use repository configured login plugin application, rather than a hard-coded programmatic one.
I've DELIBERATELY removed the skin attribute from the programmatic
configuration of the login plugin in the JavaConfigService.  This because it
should be VERY noticeable that not the proper login box as specified by the
repository based configuration of the login application is used.  Only then a
disfunctional configuration of the login application will be noticed.

By default, with a working repository, the configuration is used as specified
in the repository.  The programmatic configuration is used only as a fall
back, otherwise there is no application at all to show.  The alternative is to
show a "not available" application or to remove it all together (getting a
error status code from the application container).  But letting the
programmatic configuration of the login behave just like the repository
configuration is NOT AN OPTION.


svn path=/hippo-ecm/trunk/; revision=16136