HREPTWO-601: add faceted date add-on.
The hippostd:date node type is the normal node type to use to
house a date which will be seperated into its facet, however the
hippostd:datefields can be used to indicate a custom date field
needs to be split up into facets.  This however requires that you
add the proper derived definitions yourself.  See the provided
unit tests on the details for this.

svn path=/hippo-ecm/trunk/; revision=16147