HREPTWO-1648: initial jcr observation
Implemented a jcr observation manager that uses weak references to remove
listeners that are no longer relevant.  When a page is deserialized, the
listeners are re-registered for the newly created objects.

This solution should work in a clustered environment (i.e. not necessarily
sticky sessions) and any type of page store.

The request cycle has been altered; both ajax and non-ajax requests
will process the events that have occurred.  The render() method of the
IRenderService can now also be called in a non-ajax request.

svn path=/hippo-ecm/trunk/; revision=16153