HREPTWO-2221, HREPTWO-402, HREPTWO-1434, HREPTWO-473:
Implemented observation in a new treemodel, the CachedTreeModel.  It is an
observer as well as a (swing) model for the tree.  The tree-model has also
lost its rendering capabilities.  Rendering (tree) nodes is the responsibility
of the view, not the model.

Moved the plugin-request-target to a more natural location, as well as
reverting to the old behavior of intercepting all respond listeners.  This
is necessary to be able to filter out components that have been added to
the request target by these listeners.

Removed a lot of HippoNode usage; that something is a hippo-node should only
be relevant to a few plugins.  By making the JcrNodeModel return a strict
jcr node, adherence to this rule is hopefully encouraged.  (and, furthermore,
atm the nodes in the version history are actually not hippo nodes)

svn path=/hippo-ecm/trunk/; revision=16194