HREPTWO-1434: implement in-session observation
In-session events are generated by the use of Session.pendingChanges().  A
simple mechanism to prevent the multiple firing of the same event has been
implemented.

The wicket tree did not stand up to handling multiple events in a single
request; TreeItem children are now built eagerly iso waiting for the
ITargetRespondListener to be invoked.  Only top-level items are added to
the request target, i.e. child items are filtered out.

The tree node lookup mechanism has been modified to return TreePath instances.
The prefetching has been removed from the lookup path, as this might lead to
many nodes being retrieved that won't be used again.

svn path=/hippo-ecm/trunk/; revision=16294