HREPTWO-2244: use configured client-class-name
And move all plugins that are on the dashboard to the separate dashboard
cluster.  Removed the initialized optimization; always re-render the
wireframe layout.  The UnitBehavior makes sure that this happens by adding
the component that contains the wireframe to the ajax request target.  (in
non-ajax requests, this is not necessary)

svn path=/hippo-ecm/trunk/; revision=16417