HREPTWO-2279: initial support for faceted browsing in CMS
To be able to create a faceted tree, it is now possible to add mixins
dynamically to documents.  In the defaultcontent addon, the
hippostd:taggable mixin can be set in this way.

Created template for the hippostd:taggable mixin.  FieldPlugin now also
supports mixins; when "type" is set, it is used for field lookup iso the
default (primary) type of a model.

The FacetSearchProvider filters the number of documents to provide a view
where only one representative is in the list.

svn path=/hippo-ecm/trunk/; revision=16998