HREPTWO-1827: re-enabling the "create namespace" workflow step
A separate thread is started in the repository to refresh initialization
data.  This is necessary since all event listeners receive their events
synchronously.  Since initialization nodes are changed during this
processing, synchronous refresh leads to a concurrent modification exception
on the save() that triggered the refresh in the first place!

In the frontend code, configuration for the context menu has been added to
all tree views.  Furthermore, some i18n changes to get a usable dialog.

svn path=/hippo-ecm/trunk/; revision=17343