HREPTWO-2267: "ocm" listeners
Introducing the JcrObject, which can be used to implement the old listener
pattern when used in conjunction with the IServiceFactory.  Subclasses can
implement java-bean style listeners by translating jcr events into business
events.  Instances of these objects are handed out by the service factory,
i.e. they are bound to the context that requested them.  This means that
listeners can be managed by the framework; when a plugin is stopped, the
business object instance is no longer referenced so its OK when there are
remaining listeners.

The service factory is an interface that is recognised by the framework.  It
can be registered under a particular service name, and is called with its
getService method when a different plugin wants to acquire the service.

Two of the stores have been mapped to a standard interface, IStore<T>, that
can be used to create, find and delete business objects.  E.g.
ITypeDescriptor objects.  The JcrTypeStore and JcrPrototypeStore have been
moved to the only location where they are used, i.e. the editor.

The template editor workflow configuration has been moved to the editor
addon, under the "editor" category.  Translations for this workflow were
added to the current activity plugin.  Note: this plugin should become
independent of workflow impl's at some point.

The template builder has been split into "cluster editing" and "type
editing" parts.  The template editor helper provides the neccessary
behavior (e.g. remove fields when plugin is removed).  At the moment,
the editor class for a plugin is hardcoded in the preview cluster config.
TODO: establish some kind of convention (or perhaps an annotation) to find
this editor class.

The ListenerList utility class has been introduced to safely iterate over a
list of listeners.  Listeners may come and go as other listeners are
notified, which leads to concurrent modification errors with a standard
List.

svn path=/hippo-ecm/trunk/; revision=17435