HREPTWO-2346: intercept click events at body
This eliminates lookups up the hierarchy and the counters that were used to
detect when menus should actually be closed.

The new behaviors EventStoppingBehavior and EventStoppingDecorator call
Wicket.stopEvent to stop the event from bubbling up the hierarchy.

The ContextMenuBehavior, with its associated contextmenu.js, maintain a
list of active context menus on the client; the ajax call is only made if
the menu (corresponding to the component that the behavior is attached to)
is in this list.

svn path=/hippo-ecm/trunk/; revision=17590