HREPTWO-2443:
Plugin that is being edited is explicitly started by its editor plugin.  This
should make it possible to have fine grained ajax updates.

Plugin config now maintains a list of child configs.  This is necessary to
maintain the listener list.  Similarly, the JcrConfigService maintains a cache
of clusters for each plugin that uses it.

The render services that are created with the RenderPlugin.newPlugin method are
called in the render phase.

svn path=/hippo-ecm/trunk/; revision=17604