HREPTWO-2435: allow observation of plugin config in preview
The ClusterConfigDecorator has been split into a generic abstract decorator
and a specific use-case.  (variable expansion)  The abstract decorators are
immutable (put() throws unsupported).

The TemplateBuilder no longer updates the complete template when the
configuration of one of the plugins has changed.  This is left to the
corresponding editor plugin.

svn path=/hippo-ecm/trunk/; revision=17610