HREPTWO-2424: restructure focussing of browser
Focussing the "browser" is now accomplished by requesting focus(null) on the
IRenderService that is available on the same service name as the browser
service.  Thus, whereas the browser service itself makes sure that the correct
node is selected, the associated render service gives clients the possibility
to switch focus to the _visual_ browser.

svn path=/hippo-ecm/trunk/; revision=17657