HREPTWO-2531: multiple previews
Implement multiple previews, making it possible to have editors open and switch to/from preview.  Two types of editor are currently available, the default editor and the hippostdpublishable editor.

The helper class ServiceContext, makes it straightforward to safely start and stop services from a plugin.  All trackers and services that are registered by the service will be removed when the service context is stopped.

svn path=/hippo-ecm/trunk/; revision=18375