HREPTWO-2432: reenable facet search observer
Events are generated for facetsearches that exist at the time of the first listener registration.  I.e. new facetsearches are only visible after a logout/login cycle.

The strong reference to the session has been replaced by a weak reference, to give the garbage collector a fair chance of collecting it.

svn path=/hippo-ecm/trunk/; revision=18660