HREPTWO-2797: couple JcrObject to observations
Eliminate the listeners; for users that really want to, they can build listeners on top of the events.

The main problem that is addressed by this fix is that listeners were directly registered with objects.  To clean up these listeners would require an additional observation indirection layer, whereas sending events to observers is an easy existing mechanism.

svn path=/hippo-ecm/trunk/; revision=18726