HREPTWO-2783: show compound types in field list
The editable types are now a query instead of a list in the template list configuration.
Furthermore, the builder blocks are no longer shown when previewing a template.
The typedescriptors have been fitted with a new method isType() to facilitate determining whether one type is a sub-type of the other.  The TypeLocator makes this possible by wrapping multiple type descriptor stores.

svn path=/hippo-ecm/trunk/; revision=18745