HREPTWO-2789, HREPTWO-2816: repositioning
it is now possible to select a layout for a document type, use a two-column layout and a single-row layout.  Each variant of a layout can have its own description.
The target list, where a new field will be added, is selectable though this is not yet reflected in the UI.  (i.e. one has to try to add a field to find out)

svn path=/hippo-ecm/trunk/; revision=18854