HREPTWO-2849: drop events when session is no longer live + null-check in plugin
The problem was that when events were processed, this would lead to calls in the session that was used to register the listener.  When running update-all-content, the session is closed and a new one is opened.  So none of the events that were queued in the listener could be processed.

The null-pointer from the TemplateTypeEditorPlugin was most likely due to a missing template.  This error also occurs when selecting the Document template in the defaultcontent namespace.

svn path=/hippo-ecm/trunk/; revision=18914