HREPTWO-2636, HREPTWO-2560, HREPTWO-2635, HREPTWO-1887: give feedback in editor
The hack that displayed the "document saved" message has been removed, as it interfered with the other feedback panels.

The document save message has become part of the editing reviewed-actions plugin.
Validator services are called before any workflow step is executed.

The MenuBar menu item list has been modified to use the item index as their id.  This makes the menu a lot more robust, i.e. it can handle events even when it has been redrawn.  This happens quite frequently when clicking "save" or "done" with a pending change in an input field.

svn path=/hippo-ecm/trunk/; revision=19035