HREPTWO-2973: don't rely on observation on handle
The problem was that, with a published variant present, the draft will be
turned into the unpublished by setting a property.  No changes on the handle
node are therefore necessary.

Solution is to update the editor when it is validated.  This occurs in each
request, so may be more expensive than a listener, but it is certainly
reliable.


svn path=/hippo-ecm/trunk/; revision=19339