HREPTWO-2977: restore facet(sub)search folder view
Facet (sub) search nodes don't update until a session.refresh(false) or session.getRootNode().refresh(false) is invoked.  This does happen when a workflow step is invoked, so the session that executes the workflow step will get to see the changes.  Other sessions don't get to see the changes until they refresh their own session.

Refreshing a session is an expensive operation, because it throws away all the session state.  This is especially expensive when running in SPI mode.  So I haven't implemented a session refresh when an facet-search-event is generated.

svn path=/hippo-ecm/trunk/; revision=19373