HREPTWO-2912: fix cluster config - type descriptor synchronization in template builder
When a field was added and removed in the same edit cycle, the field descriptor remained in the type and the property in the prototype.  These are now updated correctly.

The JcrObject OCM base class has been made abstract, as it is not intended to be used directly.  Its processEvents method has been made abstract to enforce suitable event generation in subclasses.

svn path=/hippo-ecm/trunk/; revision=19407