HREPTWO-2794: implement value and child node reordering
Reordering is implemented for child nodes and property values.  Property
value reordering uses the observation mechanism to notify plugins about the
change.  Child node reordering does not generate events; creating those from
pendingChanges is too expensive, as it requires expanding child nodes when
they are not referenceable.

The ValueTemplateProvider is now able to convert a single-valued property
into a multi-valued one.  This was necessary to have a fluent template
building experience, but might be better fixed in the builder itself.

The JcrItemModel now suppresses some InvalidItemStateExceptions; these are
the result of the item having been removed in the session.


svn path=/hippo-ecm/trunk/; revision=19451