HREPTWO-2794: indicate when types have been created or modified
When a type has been modified or a new one has been created, a state is shown that is similar to the one shown for documents.

The remodel dialog has been modified to do a non-ajax close, so that the page is rendered completely.  This is necessary, as the JCR session has been closed and reopened.  No events therefore exist to trigger appropriate updates of the page.

svn path=/hippo-ecm/trunk/; revision=19460