HREPTWO-2275: more careful closing of the dialog
After a request has been handled with "close dialog" return javascript, other requests may still come in on components in the dialog.  Client-side events may still bubble before the close javascript executes.

When the initial request is non-ajax, no close javascript can be sent.  The window close callback is invoked and the dialog is not re-rendered.

svn path=/hippo-ecm/trunk/; revision=19677