HSTTWO-840: Improving log message to print window name and component class name. Also, commenting out test dependency on hippo-ecm-repository-engine because its transitive dependencies on quartz makes build slow. Later we can probably restore this dependency if we can use local repository for unit tests.
By the way, any exception occurred during render phase in jsp page cannot be logged because JSP engine of Tomcat6 does not propagate the exceptions or errors outside.
So, for that kind of jsp errors, developers should look into the container-provided log files.

Anyway, thanks to Jeroen Hoffman for this pointer!

svn path=/ecm/site-toolkit/trunk/; revision=19694