HREPTWO-3118: listen to folder model service in picker cluster
There were a few problems.  The compatibility workflow plugin created its own model reference, that was not used by the plugins in the cluster.  It now listens to the model reference that is created by the cluster.  The other problem was that the workflows plugin instantiated workflow plugins with the same id, causing plugins to be stopped but still being shown in the menu.

svn path=/hippo-ecm/trunk/; revision=19867