HREPTWO-3141: separate editor type for templatetype nodes
Since update-all-content changes the types and prototypes, the document type editor has to switch prototype.  When an editor is still open for a document type, it is now closed when update-all-content runs.  Furthermore, a number of detach() calls have been added.  The dialog window also removes the link to the title.

The plugin context throws a runtime exception when it is being serialized after having been stopped.  In this case, some plugin keeps a reference to a service or a model that was made available by the plugin that was stopped.  This is not legal; plugins may not hold on to such references unless they handle the appearance/disappearance of services themselves.

svn path=/hippo-ecm/trunk/; revision=19928