HREPTWO-3280: don't open root node in editor
The problem was that parsing the event resulted in a model for the root node.  Browsing to that made the root node the active browser document model, causing the editor to open it.  The browseservice now deals with selecting the root node; the browse link target now understands the version history (partially), so that the link is no longer to the root node.  Instead, it points to the handle that corresponds to the document that was restored.

The JcrObject base class was modified to throw an exception when no Node is present.  This should eliminate many null-checks in favor of exception handling that is likely to be needed anyway.

svn path=/hippo-ecm/trunk/; revision=20207