HREPTWO-3140: support more flexible type names
Type names can now contain lower and uppercase ASCII characters (a-z, A-Z), underscore (_) and hyphen (-).
It is not possible to give the same freedom as in document names, as developers will have to deal with the encoded variants of these names directly.  Spaces and colons, for instance, will be encoded into a _x00??_ format.

svn path=/hippo-ecm/trunk/; revision=20238