HREPTWO-3396: indicate which field is invalid
The validation results from the validation engine are filtered by the field plugins.  When a template that is used by a field plugin does not have the validation.model reference, or the field is single-valued, then the field plugin handles the invalid indication itself.  Otherwise, this is left to the template itself.

svn path=/hippo-ecm/trunk/; revision=20701